image-manipulator/pyproject.toml
[project]
name = "image-manipulator"
version = "0.1.0"
description = "Package with libraries to manipulate images"
authors = [
    {name = "Siddhant Misra"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
isort = "^6.0.1"

